Clean up state, remove Portal edits
